+++ pyproject.toml
@@ -1,0 +1,18 @@
+[tool.poetry]
+name = "tesspoint"
+version = "1.0.0"
+description = "High Precision TESS pointing tool"
+authors = ["Your Name <you@example.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+numpy = "^1.23.0"
+
+[tool.poetry.dev-dependencies]
+pytest = "^7.1.2"
+jupyterlab = "^3.4.3"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
