--- src/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import  # noqa
-
-import os  # noqa
-
-PACKAGEDIR = os.path.abspath(os.path.dirname(__file__))
