+++ src/tesspoint/data/pointings.csv
@@ -1,0 +1,70 @@
+sector,ra,dec,roll
+1,352.6844,-64.8531,222.1532
+2,16.5571,-54.016000000000005,220.4335
+3,36.3138,-44.25899999999999,213.0384
+4,55.007,-36.641999999999996,202.8302
+5,73.5382,-31.9349,191.0517
+6,92.0096,-30.5839,178.6367
+7,110.2559,-32.6344,166.4476
+8,128.1156,-37.736999999999995,155.3091
+9,145.9071,-45.3044,145.9163
+10,165.0475,-54.8165,139.1724
+11,189.1247,-65.5369,138.0761
+12,229.5885,-75.1256,153.9773
+13,298.6671,-76.3281,198.9378
+14,276.7169,62.4756,32.2329
+15,280.3985,64.0671,55.4277
+16,282.4427,66.1422,79.4699
+17,351.2381,57.8456,41.9686
+18,16.1103,67.9575,40.5453
+19,60.2026,76.2343,19.6463
+20,129.3867,75.252,334.5689
+21,171.7951,65.1924,317.9495
+22,197.1008,53.7434,319.6992
+23,217.2879,43.8074,327.4246
+24,261.4516,63.1181,317.2624
+25,265.6098,61.9383,339.5293
+26,270.1381,61.5637,0.6038
+27,326.8525,-72.4265,214.5061
+28,357.2944,-63.0056,222.5216
+29,18.919,-52.8296,219.797
+30,38.3564,-43.3178,212.0441
+31,57.6357,-35.7835,201.2334
+32,77.1891,-31.3957,188.6263
+33,96.5996,-30.7848,175.5369
+34,115.2951,-33.779,163.1916
+35,133.2035,-39.6871,152.4006
+36,150.9497,-47.7512,143.7306
+37,170.25400000000002,-57.3725,138.1685
+38,195.7176,-67.8307,139.3519
+39,242.1981,-76.3969,161.5986
+40,273.0766,61.745,14.1539
+41,277.6209,62.763999999999996,37.2224
+42,13.014000000000001,6.3337,292.8009
+43,49.526,18.9737,284.9617
+44,89.6066,24.1343,270.1557
+45,130.296,19.0181,255.0927
+46,157.6997,10.0922,248.4063
+47,143.3807,73.1125,327.10200000000003
+48,179.4254,62.1038,317.4166
+49,202.6424,50.9532,321.3516
+50,221.8575,41.7577,329.73400000000004
+51,239.4257,35.2333,339.865
+52,266.3618,61.818999999999996,343.1429
+53,270.8126,61.5761,3.6838
+54,290.121,32.6073,13.4565
+55,307.8655,37.6464,24.5369
+56,324.2778,46.3448,36.2524
+57,344.2275,56.4121,44.01
+58,9.3118,67.6524,45.3615
+59,52.9755,77.1746,26.5121
+60,125.6742,77.3113,337.3244
+61,118.0446,-36.0902,162.2198
+62,135.2412,-42.2415,151.5884
+63,153.0613,-50.6996,142.7405
+64,173.2653,-60.865,137.281
+65,201.6239,-71.5724,140.7443
+66,259.1702,-78.7974,173.9147
+67,326.7691,-74.2796,217.4678
+68,359.2829,-64.2357,226.0975
+69,20.0449,-54.2315,222.7721
